Fix: fixes calculation bug for size analisys
